+++ packages/sdk/browser/__tests__/platform/BrowserInfo.test.ts
@@ -1,5 +1,7 @@
 import BrowserInfo from '../../src/platform/BrowserInfo';
 
+const packageJson = require('../../package.json');
+
 it('returns correct platform data', () => {
   const browserInfo = new BrowserInfo({});
   expect(browserInfo.platformData()).toEqual({
@@ -10,8 +12,8 @@ it('returns correct platform data', () => {
 it('returns correct SDK data without wrapper info', () => {
   const browserInfo = new BrowserInfo({});
   expect(browserInfo.sdkData()).toEqual({
-    name: '@launchdarkly/js-client-sdk',
-    version: '0.0.0',
+    name: packageJson.name,
+    version: packageJson.version,
     userAgentBase: 'JSClient',
   });
 });
@@ -19,8 +21,8 @@ it('returns correct SDK data without wrapper info', () => {
 it('returns correct SDK data with wrapper name', () => {
   const browserInfo = new BrowserInfo({ wrapperName: 'test-wrapper' });
   expect(browserInfo.sdkData()).toEqual({
-    name: '@launchdarkly/js-client-sdk',
-    version: '0.0.0',
+    name: packageJson.name,
+    version: packageJson.version,
     userAgentBase: 'JSClient',
     wrapperName: 'test-wrapper',
   });
@@ -29,8 +31,8 @@ it('returns correct SDK data with wrapper name', () => {
 it('returns correct SDK data with wrapper version', () => {
   const browserInfo = new BrowserInfo({ wrapperVersion: '1.0.0' });
   expect(browserInfo.sdkData()).toEqual({
-    name: '@launchdarkly/js-client-sdk',
-    version: '0.0.0',
+    name: packageJson.name,
+    version: packageJson.version,
     userAgentBase: 'JSClient',
     wrapperVersion: '1.0.0',
   });
@@ -42,8 +44,8 @@ it('returns correct SDK data with both wrapper name and version', () => {
     wrapperVersion: '1.0.0',
   });
   expect(browserInfo.sdkData()).toEqual({
-    name: '@launchdarkly/js-client-sdk',
-    version: '0.0.0',
+    name: packageJson.name,
+    version: packageJson.version,
     userAgentBase: 'JSClient',
     wrapperName: 'test-wrapper',
     wrapperVersion: '1.0.0',