+++ packages/sdk/server-ai/src/sdkInfo.ts
@@ -1,2 +1,2 @@
 export const aiSdkName = '@launchdarkly/server-sdk-ai';
-export const aiSdkVersion = '0.14.1'; // x-release-please-version
+export const aiSdkVersion = '0.15.0'; // x-release-please-version