+++ packages/telemetry/browser-telemetry/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.0.12](https://github.com/launchdarkly/js-core/compare/browser-telemetry-v1.0.11...browser-telemetry-v1.0.12) (2025-11-21)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @launchdarkly/js-client-sdk bumped from 0.8.1 to 0.9.0
+
 ## [1.0.11](https://github.com/launchdarkly/js-core/compare/browser-telemetry-v1.0.10...browser-telemetry-v1.0.11) (2025-07-23)